+++ demo/src/App.jsx
@@ -6,12 +6,15 @@ import styles from './styles.module.css'
 
 import Minimal from './sandboxes/leva-minimal/src/App'
 import PluginSpring from './sandboxes/leva-plugin-spring/src/App'
+import PluginPlot from './sandboxes/leva-plugin-plot/src/App'
+import PluginBezier from './sandboxes/leva-plugin-bezier/src/App'
 import Busy from './sandboxes/leva-busy/src/App'
 import AdvancedPanels from './sandboxes/leva-advanced-panels/src/App'
 import Scroll from './sandboxes/leva-scroll/src/App'
 import UI from './sandboxes/leva-ui/src/App'
 import Theme from './sandboxes/leva-theme/src/App'
 import CustomPlugin from './sandboxes/leva-custom-plugin/src/App'
+import LevaTransient from './sandboxes/leva-transient/src/App'
 
 const { styled } = createCss({
   theme: {
@@ -30,12 +33,15 @@ const Page = styled('div', {
 
 const links = {
   'leva-minimal': Minimal,
-  'leva-plugin-spring': PluginSpring,
   'leva-busy': Busy,
   'leva-advanced-panels': AdvancedPanels,
   'leva-scroll': Scroll,
   'leva-ui': UI,
   'leva-theme': Theme,
+  'leva-transient': LevaTransient,
+  'leva-plugin-spring': PluginSpring,
+  'leva-plugin-plot': PluginPlot,
+  'leva-plugin-bezier': PluginBezier,
   'leva-custom-plugin': CustomPlugin,
 }