+++ packages/leva/src/components/Control/ControlInput.tsx
@@ -1,26 +1,21 @@
 import React from 'react'
 import { Plugins } from '../../plugin'
-import { log, LevaErrors } from '../../utils/log'
+import { warn, LevaErrors } from '../../utils/log'
 import { InputContext } from '../../context'
-import { useValue } from '../../hooks'
+import { useInputSetters } from '../../hooks'
 import { StyledInputWrapper } from '../UI/StyledUI'
+import type { DataInput } from '../../types'
 
-type ControlInputProps<V, Settings extends object> = {
-  type: string
-  label: string
+type ControlInputProps = Omit<DataInput, '__refCount' | 'key'> & {
   valueKey: string
   path: string
-  value: V
-  settings: Settings
-  disabled: boolean
+  storeId: string
   setValue: (value: any) => void
   setSettings: (settings: any) => void
-  optional: boolean
-  hint?: string
   disable: (flag: boolean) => void
 }
 
-export function ControlInput<V, Settings extends object>({
+export function ControlInput({
   type,
   label,
   path,
@@ -30,12 +25,12 @@ export function ControlInput<V, Settings extends object>({
   setValue,
   disabled,
   ...rest
-}: ControlInputProps<V, Settings>) {
-  const { displayValue, onChange, onUpdate } = useValue({ type, value, settings, setValue })
+}: ControlInputProps) {
+  const { displayValue, onChange, onUpdate } = useInputSetters({ type, value, settings, setValue })
 
   const Input = Plugins[type].component
   if (!Input) {
-    log(LevaErrors.NO_COMPONENT_FOR_TYPE, type, path)
+    warn(LevaErrors.NO_COMPONENT_FOR_TYPE, type, path)
     return null
   }