+++ packages/leva/src/components/Select/Select.tsx
@@ -1,37 +1,46 @@
-import React from 'react'
+import React, { useRef } from 'react'
 import { useInputContext } from '../../context'
 import { Label, Row, Chevron } from '../UI'
-import { StyledSelect, SelectContainer } from './StyledSelect'
+import { NativeSelect, PresentationalSelect, SelectContainer } from './StyledSelect'
 import type { SelectProps } from './select-types'
 
-function Select({
+export function Select({
   displayValue,
+  value,
   onUpdate,
   id,
   settings,
-}: Pick<SelectProps, 'displayValue' | 'onUpdate' | 'id' | 'settings'>) {
-  const { keys, values } = settings!
+}: Pick<SelectProps, 'value' | 'displayValue' | 'onUpdate' | 'id' | 'settings'>) {
+  const { keys, values } = settings
+  const lastDisplayedValue = useRef<any>()
+
+  // in case the value isn't present in values (possibly when changing options
+  // via deps), remember the last correct display value.
+  if (value === values[displayValue]) {
+    lastDisplayedValue.current = keys[displayValue]
+  }
 
   return (
     <SelectContainer>
-      <StyledSelect id={id} value={displayValue} onChange={(e) => onUpdate(values[+e.currentTarget.value])}>
+      <NativeSelect id={id} value={displayValue} onChange={(e) => onUpdate(values[Number(e.currentTarget.value)])}>
         {keys.map((key, index) => (
           <option key={key} value={index}>
             {key}
           </option>
         ))}
-      </StyledSelect>
+      </NativeSelect>
+      <PresentationalSelect>{lastDisplayedValue.current}</PresentationalSelect>
       <Chevron toggled />
     </SelectContainer>
   )
 }
 
 export function SelectComponent() {
-  const { label, displayValue, onUpdate, id, settings } = useInputContext<SelectProps>()
+  const { label, value, displayValue, onUpdate, id, settings } = useInputContext<SelectProps>()
   return (
     <Row input>
       <Label>{label}</Label>
-      <Select id={id} displayValue={displayValue} onUpdate={onUpdate} settings={settings} />
+      <Select id={id} value={value} displayValue={displayValue} onUpdate={onUpdate} settings={settings} />
     </Row>
   )
 }