+++ packages/leva/src/components/Select/select-plugin.ts
@@ -1,5 +1,5 @@
 import v8n from 'v8n'
-import type { InternalSelectSettings, SelectInput } from './select-types'
+import type { SelectInput, InternalSelectSettings } from './select-types'
 
 // the options attribute is either an key value object or an array
 export const schema = (_o: any, s: any) =>
@@ -8,6 +8,11 @@ export const schema = (_o: any, s: any) =>
       options: v8n().passesAnyOf(v8n().object(), v8n().array()),
     })
     .test(s)
+
+export const sanitize = (value: any, { values }: InternalSelectSettings) => {
+  if (values.indexOf(value) < 0) throw Error(`Selected value doesn't match Select options`)
+  return value
+}
 
 export const format = (value: any, { values }: InternalSelectSettings) => {
   return values.indexOf(value)
@@ -20,7 +25,7 @@ export const normalize = (input: SelectInput) => {
 
   if (Array.isArray(options)) {
     values = options
-    keys = options.map(String)
+    keys = options.map((o) => String(o))
   } else {
     values = Object.values(options)
     keys = Object.keys(options)