+++ packages/leva/src/useControls.ts
@@ -1,14 +1,22 @@
-import { useEffect, useMemo, useCallback, useState } from 'react'
+import { useEffect, useMemo, useCallback, useState, useRef } from 'react'
 import { levaStore } from './store'
 import { folder } from './helpers'
-import { useShallowMemo, useValuesForPath } from './hooks'
+import { useDeepMemo, useValuesForPath } from './hooks'
 import { useRenderRoot } from './components/Leva'
 import type { FolderSettings, Schema, SchemaToValues, StoreType } from './types'
+import shallow from 'zustand/shallow'
 
 type HookSettings = { store?: StoreType }
 type SchemaOrFn<S extends Schema = Schema> = S | (() => S)
 
-type FunctionReturnType<S extends Schema> = [SchemaToValues<S>, (value: Partial<SchemaToValues<S>>) => void]
+type FunctionReturnType<S extends Schema> = [
+  SchemaToValues<S>,
+  (
+    value: {
+      [K in keyof Partial<SchemaToValues<S, true>>]: any
+    }
+  ) => void
+]
 
 type ReturnType<F extends SchemaOrFn> = F extends SchemaOrFn<infer S>
   ? F extends Function
@@ -92,9 +100,16 @@ export function useControls<S extends Schema, F extends SchemaOrFn<S> | string, 
 
   const schemaIsFunction = typeof schema === 'function'
 
+  // Keep track of deps to see if they changed and if there's need to recompute.
+  const depsChanged = useRef(false)
+  // We will only override the store settings and options when deps have changed
+  // and it isn't the first render
+  const firstRender = useRef(true)
+
   // Since the schema object would change on every render, we let the user have
   // control over when it should trigger a reset of the hook inputs.
-  const _schema = useShallowMemo(() => {
+  const _schema = useDeepMemo(() => {
+    depsChanged.current = true
     const s = typeof schema === 'function' ? schema() : schema
     return folderName ? { [folderName]: folder(s, folderSettings) } : s
   }, deps)
@@ -114,9 +129,20 @@ export function useControls<S extends Schema, F extends SchemaOrFn<S> | string, 
    * parses the schema inside nested folder.
    */
   const [initialData, mappedPaths] = useMemo(() => store.getDataFromSchema(_schema), [store, _schema])
+  const [allPaths, renderPaths, onChangePaths] = useMemo(() => {
+    const allPaths: string[] = []
+    const renderPaths: string[] = []
+    const onChangePaths: Record<string, (v: any) => void> = {}
+    Object.values(mappedPaths).forEach(({ path, onChange }) => {
+      allPaths.push(path)
+      if (!!onChange) onChangePaths[path] = onChange
+      else renderPaths.push(path)
+    })
+    return [allPaths, renderPaths, onChangePaths]
+  }, [mappedPaths])
 
   // Extracts the paths from the initialData and ensures order of paths.
-  const paths = useMemo(() => store.orderPaths(Object.values(mappedPaths)), [mappedPaths, store])
+  const paths = useMemo(() => store.orderPaths(allPaths), [allPaths, store])
 
   /**
    * Reactive hook returning the values from the store at given paths.
@@ -126,12 +152,14 @@ export function useControls<S extends Schema, F extends SchemaOrFn<S> | string, 
    * initalData is going to be returned on the first render. Subsequent renders
    * will call the store data.
    * */
-  const values = useValuesForPath(store, paths, initialData)
+  const values = useValuesForPath(store, renderPaths, initialData)
 
   const set = useCallback(
     (values: Record<string, any>) => {
-      // @ts-ignore
-      const _values = Object.entries(values).reduce((acc, [p, v]) => Object.assign(acc, { [mappedPaths[p]]: v }), {})
+      const _values = Object.entries(values).reduce(
+        (acc, [p, v]) => Object.assign(acc, { [mappedPaths[p].path]: v }),
+        {}
+      )
       store.set(_values)
     },
     [store, mappedPaths]
@@ -142,9 +170,27 @@ export function useControls<S extends Schema, F extends SchemaOrFn<S> | string, 
     // Note that doing this while rendering (ie in useMemo) would make
     // things easier and remove the need for initializing useValuesForPath but
     // it breaks the ref from Monitor.
-    store.addData(initialData)
+
+    // we override the settings when deps have changed and this isn't the first
+    // render
+    const shouldOverrideSettings = !firstRender.current && depsChanged.current
+    store.addData(initialData, shouldOverrideSettings)
+    firstRender.current = false
+    depsChanged.current = false
     return () => store.disposePaths(paths)
   }, [store, paths, initialData])
+
+  useEffect(() => {
+    // let's handle transient subscriptions
+    const unsubscriptions: (() => void)[] = []
+    Object.entries(onChangePaths).forEach(([path, onChange]) => {
+      onChange(store.get(path))
+      // @ts-ignore
+      const unsub = store.useStore.subscribe(onChange, (s) => s.data[path].value, shallow)
+      unsubscriptions.push(unsub)
+    })
+    return () => unsubscriptions.forEach((unsub) => unsub())
+  }, [store, onChangePaths])
 
   if (schemaIsFunction) return [values, set] as any
   return values as any